pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secgate"
version = "1.0.0"
description = "Linux 服务器安全网关管理平台"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "flask>=3.0",
    "psutil>=5.9",
    "requests>=2.28",
    "itsdangerous>=2.0",
    "pyyaml>=6.0",
    "gunicorn>=21.0",
]

[project.optional-dependencies]
agent = ["chainlit>=2.0"]

[project.scripts]
secgate = "secgate_pkg.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
secgate_pkg = ["project/**/*"]
